+++ src/types/client/audio.ts
@@ -4,3 +4,8 @@ export type SearchAudioResult = {
   count: number;
   audios: AudioItem[];
 };
+
+export type AddAudioResult = {
+  audioId: number;
+  ownerId: number;
+};